dynamic single-feature plots optimized
