+++ src/components/Differential/PlotlySingleFeature.jsx
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
-// import { Dimmer, Loader } from 'semantic-ui-react';
 import Plot from 'react-plotly.js';
-import { reviseLayout, clickDownload } from '../Shared/helpers';
+import { reviseLayout, clickDownload, loadingDimmer } from '../Shared/helpers';
 import '../Shared/PlotlyOverrides.scss';
 
 export default class PlotlySingleFeature extends Component {
@@ -13,14 +12,7 @@ export default class PlotlySingleFeature extends Component {
     loading: true,
   };
   componentDidMount() {
-    const json = this.getJson();
-    // console.log("mounted");
-    this.setState({
-      json: {
-        data: json.data,
-        layout: json.layout,
-      },
-    });
+    this.getJson();
   }
 
   componentDidUpdate(prevProps) {
@@ -30,81 +22,30 @@ export default class PlotlySingleFeature extends Component {
     ) {
       clickDownload(this.props.parentNode);
     }
-    if (
-      prevProps.plotKey !== this.props.plotKey ||
-      prevProps.plotId !== this.props.plotId ||
-      prevProps.dimensions !== this.props.dimensions ||
-      prevProps.plotlyData?.length !== this.props.plotlyData?.length
-    ) {
-      // console.log("updated!");
-      const json = this.getJson();
-      this.setState({
-        json: {
-          data: json.data,
-          layout: json.layout,
-        },
-      });
+    if (this.props.cacheString !== prevProps.cacheString) {
+      this.getJson();
     }
   }
 
-  // componentWillUnmount() {
-  //   console.log("unmounted")
-  //   this.setState({
-  //     json: {
-  //       data: null,
-  //       layout: null,
-  //     },
-  //   });
-  // }
-
-  // purgePlot = () => {
-  //   console.log("purge");
-  //   this.setState({
-  //     json: {
-  //       data: null,
-  //       layout: null,
-  //     },
-  //   });
-  // }
-
   getJson = () => {
-    const {
-      differentialStudy,
-      differentialModel,
-      differentialTest,
-      plotKey,
-      plotId,
-      plotlyData,
-      width,
-      height,
-    } = this.props;
-    const cacheKey = `PlotlySingleFeature_${width}_${height}_${differentialStudy}_${differentialModel}_${differentialTest}_${plotKey}_${plotId}`;
-    if (this[cacheKey] != null) {
-      return this[cacheKey];
-    } else {
-      const parsedData = JSON.parse(plotlyData);
-      const data = parsedData?.data || null;
-      let layout = parsedData?.layout || null;
-      if (layout) {
-        layout = reviseLayout(layout, width, height);
-      }
-      this[cacheKey] = { data, layout };
-      return { data, layout };
+    const { plotlyData, width, height } = this.props;
+    const parsedData = JSON.parse(plotlyData);
+    const data = parsedData?.data || null;
+    let layout = parsedData?.layout || null;
+    if (layout) {
+      layout = reviseLayout(layout, width, height);
     }
-  };
-
-  handleInitialized = () => {
     this.setState({
-      loading: false,
+      json: {
+        data,
+        layout,
+        loading: false,
+      },
     });
   };
 
   render() {
-    const { plotName, featuresLength, plotlyExportType } = this.props;
-    // const loader = this.state.json?.data ? null : (
-    //   <Dimmer active inverted>
-    //   <Loader size="large">Loading...</Loader>
-    // </Dimmer>
+    const { plotName, plotKey, plotlyExportType } = this.props;
     const config = {
       modeBarButtonsToRemove: ['sendDataToCloud'],
       displayModeBar: true,
@@ -112,25 +53,21 @@ export default class PlotlySingleFeature extends Component {
       displaylogo: false,
       toImageButtonOptions: {
         format: plotlyExportType, // one of png, svg, jpeg, webp
-        filename: `${plotName}_(${featuresLength}-features)`,
+        filename: `${plotName}_${plotKey}`,
       },
     };
-    // console.log(JSON.stringify(this.state.json?.data));
     return (
       <div>
-        {this.state.json?.data && this.state.json?.layout && (
+        {this.state.json.data && this.state.json.layout && (
           <Plot
             data={this.state.json.data}
             layout={this.state.json.layout}
             config={config}
-            // onPurge={this.purgePlot}
-            // onInitialized={figure => this.setState({ figure, loading: false })}
-            // onUpdate={figure => this.setState(figure)}
           />
         )}
-        {/* <span id="PlotSingleFeatureDataLoader">
-          {this.state.loading && loader}
-        </span> */}
+        <span id="PlotSingleFeatureDataLoader">
+          {!this.state.loading && loadingDimmer}
+        </span>
       </div>
     );
   }
