+++ src/components/Differential/PlotsSingleFeature.jsx
@@ -1,10 +1,7 @@
 import React, { Component } from 'react';
-import { Loader, Dimmer, Tab, Icon, Button, Dropdown } from 'semantic-ui-react';
-import SVG from 'react-inlinesvg';
-import { roundToPrecision } from '../Shared/helpers';
+import { Loader, Dimmer, Icon, Button, Dropdown } from 'semantic-ui-react';
 import ButtonActions from '../Shared/ButtonActions';
-import MetafeaturesTableDynamic from './MetafeaturesTableDynamic';
-import PlotlySingleFeature from './PlotlySingleFeature';
+import TabSingleFeature from './TabSingleFeature';
 import './PlotsDynamic.scss';
 import '../Shared/Plotly.scss';
 
@@ -18,7 +15,6 @@ class PlotsSingleFeature extends Component {
     txtFlagSFPlots: false,
     plotlyExport: false,
     plotlyExportType: 'svg',
-    isPlotlyPlot: true,
   };
 
   differentialDetailPlotsSingleFeatureRef = React.createRef();
@@ -41,7 +37,11 @@ class PlotsSingleFeature extends Component {
   setButtonVisibility = () => {
     if (this.props.differentialPlotTypes?.length > 0) {
       const isMetaFeatureTab =
-        this.metaFeaturesTableDynamicRef.current !== null ? true : false;
+        this.state.activeTabIndexPlotsSingleFeature <
+        this.props.singleFeaturePlotTypes.length
+          ? false
+          : true;
+      // this.metaFeaturesTableDynamicRef.current !== null ? true : false;
       this.setState({
         // display excel and text on meta-feature tab
         excelFlagSFPlots: isMetaFeatureTab,
@@ -51,12 +51,6 @@ class PlotsSingleFeature extends Component {
         pngFlagSFPlots: !isMetaFeatureTab,
       });
     }
-  };
-
-  handleTabChangeSingleFeature = (e, { activeIndex }) => {
-    this.setState({
-      activeTabIndexPlotsSingleFeature: activeIndex,
-    });
   };
 
   handlePlotDropdownChangeSingleFeature = (e, { value }) => {
@@ -69,127 +63,6 @@ class PlotsSingleFeature extends Component {
     const { plotSingleFeatureData } = this.props;
     const key = plotSingleFeatureData.key;
     this.props.onGetPlotTransitionRef(key, null, plotSingleFeatureData, true);
-  };
-
-  getSVGPanesSingleFeature = activeTabIndexPlotsSingleFeature => {
-    const {
-      divWidth,
-      divHeight,
-      pxToPtRatio,
-      pointSize,
-      plotSingleFeatureData,
-      modelSpecificMetaFeaturesExist,
-      differentialStudy,
-      differentialModel,
-      differentialTest,
-      plotOverlayLoaded,
-      singleFeaturePlotTypes,
-    } = this.props;
-    let panes = [];
-    let dimensions = '';
-    let divWidthPt = 0;
-    let divHeightPt = 0;
-    let divWidthPadding = 0;
-    let divHeightPadding = 0;
-    if (divWidth && divHeight && pxToPtRatio) {
-      divWidthPadding = divWidth * 0.95;
-      divHeightPadding = divHeight * 0.95 - 38;
-      divWidthPt = roundToPrecision(divWidthPadding / pxToPtRatio, 1);
-      divHeightPt = roundToPrecision(divHeightPadding / pxToPtRatio, 1);
-      const divWidthPtString = `width=${divWidthPt}`;
-      const divHeightPtString = `&height=${divHeightPt}`;
-      const pointSizeString = `&pointsize=${pointSize}`;
-      dimensions = `?${divWidthPtString}${divHeightPtString}${pointSizeString}`;
-    }
-    if (plotSingleFeatureData && plotSingleFeatureData.svg?.length) {
-      // since this call is in render, index determines the one tab to display (svg, plotly or feature data)
-      if (activeTabIndexPlotsSingleFeature < singleFeaturePlotTypes.length) {
-        const plotKey = plotSingleFeatureData.key;
-        const plotId =
-          singleFeaturePlotTypes[activeTabIndexPlotsSingleFeature].plotID;
-        // const cacheKey = `singleFeaturePanes_${dimensions}_${differentialStudy}_${differentialModel}_${differentialTest}_${plotKey}_${plotId}_${activeTabIndexPlotsSingleFeature}`;
-        // if (this[cacheKey] != null) {
-        // console.log(
-        //   `single features render cached ${cacheKey}`,
-        //   this[cacheKey],
-        // );
-        //   return this[cacheKey];
-        // } else {
-        const s = plotSingleFeatureData?.svg[activeTabIndexPlotsSingleFeature];
-        if (s) {
-          const srcUrl = `${s.svg}${dimensions}`;
-          const isPlotlyPlot = s.plotType.plotType.includes('plotly');
-          const svgPanes = {
-            menuItem: `${s.plotType.plotDisplay}`,
-            render: () => (
-              <Tab.Pane
-                attached="true"
-                as="div"
-                key={`${activeTabIndexPlotsSingleFeature}-${s.plotType.plotID}-pane-singleFeature`}
-              >
-                <div id="PlotsSingleFeatureContainer" className="svgSpan">
-                  {isPlotlyPlot ? (
-                    <PlotlySingleFeature
-                      plotlyData={s.svg}
-                      height={divHeightPadding}
-                      width={divWidthPadding}
-                      differentialStudy={differentialStudy}
-                      differentialModel={differentialModel}
-                      differentialTest={differentialTest}
-                      plotKey={plotKey}
-                      plotId={plotId}
-                      dimensions={dimensions}
-                      plotName={s.plotType.plotDisplay}
-                      plotlyExport={this.state.plotlyExport}
-                      plotlyExportType={this.state.plotlyExportType}
-                      parentNode={this.differentialDetailPlotsSingleFeatureRef}
-                    />
-                  ) : (
-                    <SVG
-                      cacheRequests={true}
-                      src={srcUrl}
-                      title={`${s.plotType.plotDisplay}`}
-                      uniqueHash={`a1f8d1-${s.plotType.plotID}-${activeTabIndexPlotsSingleFeature}`}
-                      uniquifyIDs={true}
-                    />
-                  )}
-                </div>
-              </Tab.Pane>
-            ),
-          };
-          panes = panes.concat(svgPanes);
-          // this[cacheKey] = panes;
-        }
-        // }
-      } else {
-        // if the activeTabIndex is the same as the singleFeaturePlotTypes length, it indicates app should display the Metafeatures tab
-        const isMultifeaturePlot =
-          plotSingleFeatureData?.key?.includes('features') || false;
-        if (modelSpecificMetaFeaturesExist !== false && !isMultifeaturePlot) {
-          let metafeaturesTab = [
-            {
-              menuItem: 'Feature Data',
-              render: () => (
-                <Tab.Pane attached="true" as="div">
-                  <MetafeaturesTableDynamic
-                    ref={this.metaFeaturesTableDynamicRef}
-                    differentialStudy={differentialStudy}
-                    differentialModel={differentialModel}
-                    plotOverlayLoaded={plotOverlayLoaded}
-                    plotSingleFeatureData={plotSingleFeatureData}
-                    modelSpecificMetaFeaturesExist={
-                      modelSpecificMetaFeaturesExist
-                    }
-                  />
-                </Tab.Pane>
-              ),
-            },
-          ];
-          panes = panes.concat(metafeaturesTab);
-        }
-      }
-      return panes;
-    }
   };
 
   getInstructions = () => {
@@ -253,17 +126,9 @@ class PlotsSingleFeature extends Component {
           this.props.differentialPlotTypes?.length > this.props.svgTabMax
             ? 'Show svgPlotDropdown'
             : 'Hide svgPlotDropdown';
-        const TabMenuClass =
-          this.props.differentialPlotTypes?.length > this.props.svgTabMax
-            ? 'Hide'
-            : 'Show';
         const activeTabIndexPlotsSingleFeatureVar =
           activeTabIndexPlotsSingleFeature || 0;
         const svgArray = [...plotSingleFeatureData.svg];
-        const svgPanesSingleFeature = this.getSVGPanesSingleFeature(
-          activeTabIndexPlotsSingleFeature,
-          svgArray,
-        );
         let options = [];
         options = svgArray.map(function(s, index) {
           return {
@@ -342,15 +207,12 @@ class PlotsSingleFeature extends Component {
                   : 'svgPlotDropdownSingle'
               }
             />
-            <Tab
-              menu={{
-                secondary: true,
-                pointing: true,
-                className: TabMenuClass,
-              }}
-              panes={svgPanesSingleFeature}
-              onTabChange={this.handleTabChange}
-              activeIndex={0}
+            <TabSingleFeature
+              {...this.props}
+              {...this.state}
+              differentialDetailPlotsSingleFeatureRefFwd={
+                this.differentialDetailPlotsSingleFeatureRef
+              }
             />
             <span
               className={divWidth < 450 ? 'Hide' : 'Show'}
